Work around the fact that some packages must install before configuring apt repos
We have a site-specific puppet module (profile::apt) that ensures that all
the apt repos are applied before running apt-update, and that apt-update is
run before installing any packages.  Usually, this is what you want.

However, here it wants to install some packages (apt-transport-https and
ca-certificates) that allow other signed packages to later install.  This
creates a circular dependancy with profile::apt.  This kludge tags these
modules as exempt from the packages-before-apt-update dependancy.